requirements.txt
python-dotenv
MetaTrader5
telethon
openai
